combine all steps into one, since each step output is stored in database and cause 'string or blob too big: SQLITE_TOOBIG'
